requirements.txt
paho-mqtt==1.1
pyusb==1.0.0b2
wheel==0.24.0
